feat: getRanking

+++ src/main/java/cat/itacademy/s05/t01/n01/model/Game.java
@@ -18,7 +18,7 @@ public class Game {
     @Id
     private String id;
     private String playerId;
-    private int currentScore;
+    private long currentScore;
     private List<String> cards;
     private boolean finished;
