+++ src/main/java/cat/itacademy/s05/t01/n01/model/Player.java
@@ -17,7 +17,7 @@ public class Player {
     @Id
     private int id;
     private String name;
-    private int totalScore;
+    private long totalScore;
